requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.12
ebooklib>=0.18
beautifulsoup4>=4.12.0
lxml>=5.0.0
kokoro>=0.9.4
# Orpheus TTS dependencies (optional — only needed if using engine=orpheus)
snac>=1.0.0
transformers>=4.40.0
soundfile>=0.12.1
numpy>=1.26.0
pydub>=0.25.1
httpx>=0.27.0
pyspellchecker>=0.8.0
uvicorn
